+++ libcrystfel/src/detector.h
@@ -185,7 +185,7 @@ extern double get_tt(struct image *image, double xs, double ys, int *err);
 
 extern int in_bad_region(struct detector *det, double fs, double ss);
 
-extern void record_image(struct image *image, int do_poisson, int background,
+extern void record_image(struct image *image, int do_poisson, double background,
                          gsl_rng *rng, double beam_radius, double nphotons);
 
 extern struct panel *find_panel(struct detector *det, double fs, double ss);
